mp_pytorch-0.1.1 with user guide
