+++ doc/05_nn-based_mp.md
@@ -1,0 +1,7 @@
+# 5. Neural Networks based Movement Primitives
+**The corresponding docs and demos are under construction.** 
+
+
+
+
+[Back to Overview](./)
