+++ mp_pytorch/demo/__init__.py
@@ -1,0 +1,8 @@
+from .demo_mp_config import *
+from .demo_basis_gn import *
+from .demo_data_type_cast import *
+from .demo_delay_and_scale import *
+from .demo_dmp import *
+from .demo_promp import *
+from .demo_prodmp import *
+from .demo_prodmp_autoscale import *
