+++ mp_pytorch/demo/demo_dmp.py
@@ -3,7 +3,7 @@
 """
 
 import mp_pytorch.util as util
-from demo_mp_config import get_mp_utils
+from mp_pytorch.demo import get_mp_utils
 from mp_pytorch.mp import MPFactory
